Create vaccine administration over time chart
